+++ src/components/StripePayment.vue
@@ -31,7 +31,7 @@ import { Appearance } from '@stripe/stripe-js';
 import { loadStripe } from '@stripe/stripe-js/pure';
 import { onBeforeMount, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
-import AppButton from './forms/AppButton.vue';
+import AppButton from './button/AppButton.vue';
 import MessageBox from './MessageBox.vue';
 import AppInput from './forms/AppInput.vue';
 import AppLabel from './forms/AppLabel.vue';
