+++ src/components/forms/AppImageUpload.vue
@@ -39,7 +39,7 @@ import { ref, toRef, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import axios from '../../lib/axios';
 import env from '../../env';
-import AppButton from './AppButton.vue';
+import AppButton from '../button/AppButton.vue';
 import AppLabel from './AppLabel.vue';
 
 const emit = defineEmits(['update:modelValue']);
