+++ src/components/pages/profile/account/ChangePassword.vue
@@ -52,7 +52,7 @@
 import { onBeforeMount, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { updateContact } from '../../../../utils/api/contact';
-import AppButton from '../../../forms/AppButton.vue';
+import AppButton from '../../../button/AppButton.vue';
 import MessageBox from '../../../MessageBox.vue';
 import AppInput from '../../../forms/AppInput.vue';
 import AppHeading from '../../../AppHeading.vue';
