+++ src/components/rte/RichTextEditorButton.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script lang="ts" setup>
-import AppButton from '../forms/AppButton.vue';
+import AppButton from '../button/AppButton.vue';
 
 defineProps<{
   icon: string;
