+++ src/pages/admin/callouts/index.vue
@@ -86,7 +86,7 @@ import { computed, ref, watchEffect } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRoute, useRouter } from 'vue-router';
 import { Header, SortType } from '../../../components/table/table.interface';
-import AppButton from '../../../components/forms/AppButton.vue';
+import AppButton from '../../../components/button/AppButton.vue';
 import PageTitle from '../../../components/PageTitle.vue';
 import AppTable from '../../../components/table/AppTable.vue';
 import AppAlert from '../../../components/AppAlert.vue';
