+++ src/pages/admin/contacts/add.vue
@@ -71,7 +71,7 @@ import AppHeading from '../../../components/AppHeading.vue';
 import { UpdateContribution } from '../../../components/contact/contact.interface';
 import ContactBasicFields from '../../../components/contact/ContactBasicFields.vue';
 import ContactContributionFields from '../../../components/contact/ContactContributionFields.vue';
-import AppButton from '../../../components/forms/AppButton.vue';
+import AppButton from '../../../components/button/AppButton.vue';
 import AppCheckbox from '../../../components/forms/AppCheckbox.vue';
 import AppForm from '../../../components/forms/AppForm.vue';
 import PageTitle from '../../../components/PageTitle.vue';
