+++ src/pages/auth/forgot-password.vue
@@ -55,7 +55,7 @@ meta:
 
 <script lang="ts" setup>
 import AppInput from '../../components/forms/AppInput.vue';
-import AppButton from '../../components/forms/AppButton.vue';
+import AppButton from '../../components/button/AppButton.vue';
 import { useI18n } from 'vue-i18n';
 import useVuelidate from '@vuelidate/core';
 import { reactive, ref } from 'vue';
