+++ src/pages/join/failed.vue
@@ -24,7 +24,7 @@ meta:
 
 <script lang="ts" setup>
 import JoinNotice from '../../components/pages/join/JoinNotice.vue';
-import AppButton from '../../components/forms/AppButton.vue';
+import AppButton from '../../components/button/AppButton.vue';
 import { useI18n } from 'vue-i18n';
 
 const { t } = useI18n();
