+++ src/pages/profile/index.vue
@@ -76,7 +76,7 @@ import ContributionInfo from '../../components/pages/profile/ContributionInfo.vu
 import QuickActions from '../../components/pages/profile/QuickActions.vue';
 import SectionTitle from '../../components/pages/profile/SectionTitle.vue';
 import PageTitle from '../../components/PageTitle.vue';
-import AppButton from '../../components/forms/AppButton.vue';
+import AppButton from '../../components/button/AppButton.vue';
 import AppAlert from '../../components/AppAlert.vue';
 import CalloutCard from '../../components/callout/CalloutCard.vue';
 import WelcomeMessage from '../../components/welcome-message/WelcomeMessage.vue';
